requirements.txt
python-binance
sqlalchemy
pandas
matplotlib

typing-extensions

#dev
ipython
ipdb
pytest
pylint
isort